chore(c): move C definitions to 'libbpfgo.c'
Leaving 'libbpfgo.h' only with declarations to avoid multiple
declaration errors when importing 'libbpfgo.h' from different Go files.
